+++ fast_nav/readme/nested_navigator.dart
@@ -1,0 +1,16 @@
+import 'package:fast_nav/fast_nav.dart';
+import 'package:flutter/material.dart';
+
+void example() {
+  // ...
+
+  NestedNavigator(
+    navigatorKey: GlobalKey<NavigatorState>(),
+    name: 'nested',
+    home: const Text('home'),
+  );
+
+  // ...
+
+  FastNav.push(const Text('New page'), navigatorName: 'nested');
+}