+++ fast_rx/lib/src/rx/rx_list.dart
@@ -15,12 +15,17 @@ class RxList<E> extends RxIterable<E> implements List<E> {
   }
 
   @override
+  @protected
+  List<E> get unregisteredValue {
+    return super.unregisteredValue as List<E>;
+  }
+
+  @override
   List<E> copyValue() => List.from(unregisteredValue);
 
   @override
   bool shouldNotify(Iterable<E> oldValue) =>
-      oldValue is List<E> &&
-      !listEquals(unregisteredValue as List<E>, oldValue);
+      oldValue is List<E> && !listEquals(unregisteredValue, oldValue);
 
   @override
   List<R> cast<R>() {
@@ -39,18 +44,18 @@ class RxList<E> extends RxIterable<E> implements List<E> {
 
   @override
   void operator []=(int index, E value) {
-    notifyIfChanged(() => this.value[index] = value);
+    unregisteredValue[index] = value;
   }
 
   @override
   void add(E value) {
-    this.value.add(value);
+    unregisteredValue.add(value);
     notify();
   }
 
   @override
   void addAll(Iterable<E> iterable) {
-    notifyIfChanged(() => value.addAll(iterable));
+    notifyIfChanged(() => unregisteredValue.addAll(iterable));
   }
 
   @override
@@ -60,17 +65,17 @@ class RxList<E> extends RxIterable<E> implements List<E> {
 
   @override
   void clear() {
-    notifyIfChanged(() => value.clear());
+    notifyIfChanged(() => unregisteredValue.clear());
   }
 
   @override
   void fillRange(int start, int end, [E? fillValue]) {
-    notifyIfChanged(() => value.fillRange(start, end, fillValue));
+    notifyIfChanged(() => unregisteredValue.fillRange(start, end, fillValue));
   }
 
   @override
   set first(E value) {
-    notifyIfChanged(() => this.value.first = value);
+    notifyIfChanged(() => unregisteredValue.first = value);
   }
 
   @override
@@ -90,17 +95,17 @@ class RxList<E> extends RxIterable<E> implements List<E> {
 
   @override
   void insert(int index, E element) {
-    notifyIfChanged(() => value.insert(index, element));
+    notifyIfChanged(() => unregisteredValue.insert(index, element));
   }
 
   @override
   void insertAll(int index, Iterable<E> iterable) {
-    notifyIfChanged(() => value.insertAll(index, iterable));
+    notifyIfChanged(() => unregisteredValue.insertAll(index, iterable));
   }
 
   @override
   set last(E value) {
-    notifyIfChanged(() => this.value.last = value);
+    notifyIfChanged(() => unregisteredValue.last = value);
   }
 
   @override
@@ -115,7 +120,7 @@ class RxList<E> extends RxIterable<E> implements List<E> {
 
   @override
   set length(int newLength) {
-    notifyIfChanged(() => value.length = newLength);
+    notifyIfChanged(() => unregisteredValue.length = newLength);
   }
 
   @override
@@ -137,22 +142,24 @@ class RxList<E> extends RxIterable<E> implements List<E> {
 
   @override
   void removeRange(int start, int end) {
-    notifyIfChanged(() => value.removeRange(start, end));
+    notifyIfChanged(() => unregisteredValue.removeRange(start, end));
   }
 
   @override
   void removeWhere(bool Function(E element) test) {
-    notifyIfChanged(() => value.removeWhere(test));
+    notifyIfChanged(() => unregisteredValue.removeWhere(test));
   }
 
   @override
   void replaceRange(int start, int end, Iterable<E> replacements) {
-    notifyIfChanged(() => value.replaceRange(start, end, replacements));
+    notifyIfChanged(
+      () => unregisteredValue.replaceRange(start, end, replacements),
+    );
   }
 
   @override
   void retainWhere(bool Function(E element) test) {
-    notifyIfChanged(() => value.retainWhere(test));
+    notifyIfChanged(() => unregisteredValue.retainWhere(test));
   }
 
   @override
@@ -162,23 +169,25 @@ class RxList<E> extends RxIterable<E> implements List<E> {
 
   @override
   void setAll(int index, Iterable<E> iterable) {
-    notifyIfChanged(() => value.setAll(index, iterable));
+    notifyIfChanged(() => unregisteredValue.setAll(index, iterable));
   }
 
   @override
   void setRange(int start, int end, Iterable<E> iterable, [int skipCount = 0]) {
-    notifyIfChanged(() => value.setRange(start, end, iterable, skipCount));
+    notifyIfChanged(
+      () => unregisteredValue.setRange(start, end, iterable, skipCount),
+    );
   }
 
   @override
   void shuffle([Random? random]) {
-    value.shuffle(random);
+    unregisteredValue.shuffle(random);
     notify();
   }
 
   @override
   void sort([int Function(E a, E b)? compare]) {
-    notifyIfChanged(() => value.sort(compare));
+    notifyIfChanged(() => unregisteredValue.sort(compare));
   }
 
   @override