+++ fast_rx/lib/testing.dart
@@ -1,0 +1,1 @@
+export 'testing/expect_rx_registration.dart';